refactor: reuse shared dependency resolver vite plugin in studio + app builds
